requirements.txt
flask
gunicorn
openai==0.28
python-dotenv